// convex/utils/flyio/types.ts
export interface components {
  schemas: {
      App: {
          id?: string;
          name?: string;
          organization?: components["schemas"]["Organization"];
          status?: string;
      };
      AppSecret: {
          digest?: string;
          name?: string;
          value?: string;
      };
      AppSecrets: {
          secrets?: components["schemas"]["AppSecret"][];
      };
      CheckStatus: {
          name?: string;
          output?: string;
          status?: string;
          updated_at?: string;
      };
      CreateAppRequest: {
          app_name?: string;
          enable_subdomains?: boolean;
          network?: string;
          org_slug?: string;
      };
      CreateLeaseRequest: {
          description?: string;
          /** @description seconds lease will be valid */
          ttl?: number;
      };
      CreateMachineRequest: {
          /** @description An object defining the Machine configuration */
          config?: components["schemas"]["fly.MachineConfig"];
          lease_ttl?: number;
          lsvd?: boolean;
          /** @description Unique name for this Machine. If omitted, one is generated for you */
          name?: string;
          /** @description The target region. Omitting this param launches in the same region as your WireGuard peer connection (somewhere near you). */
          region?: string;
          skip_launch?: boolean;
          skip_service_registration?: boolean;
      };
// ...
      Machine: {
          checks?: components["schemas"]["CheckStatus"][];
          config?: components["schemas"]["fly.MachineConfig"];
          created_at?: string;
          events?: components["schemas"]["MachineEvent"][];
          /** @enum {string} */
          host_status?: "ok" | "unknown" | "unreachable";
          id?: string;
          image_ref?: components["schemas"]["ImageRef"];
          incomplete_config?: components["schemas"]["fly.MachineConfig"];
          /** @description InstanceID is unique for each version of the machine */
          instance_id?: string;
          name?: string;
          /** @description Nonce is only every returned on machine creation if a lease_duration was provided. */
          nonce?: string;
          /** @description PrivateIP is the internal 6PN address of the machine. */
          private_ip?: string;
          region?: string;
          state?: string;
          updated_at?: string;
      };
// ...
      "fly.ContainerHealthcheck": {
          exec?: components["schemas"]["fly.ExecHealthcheck"];
          /** @description The number of times the check must fail before considering the container unhealthy. */
          failure_threshold?: number;
          /** @description The time in seconds to wait after a container starts before checking its health. */
          grace_period?: number;
          http?: components["schemas"]["fly.HTTPHealthcheck"];
// ...
      };
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}